+++ src/extension_core/extension.ts
@@ -35,6 +35,7 @@ export async function activate(context: vscode.ExtensionContext) {
   // I want to get semantic tokens for the current theme
   // The API is not available yet, so I'm just going to log the theme
   vscode.window.onDidChangeActiveColorTheme((e) => {viewerProvider.updateColorTheme(e);});
+  vscode.workspace.onDidChangeConfiguration((e) => {viewerProvider.updateConfiguration(e);});
 
   // #region External Commands
   context.subscriptions.push(vscode.commands.registerCommand('vaporview.openFile', (uri) => {